requirements.txt
jinja2>=3.1.3
markdown>=3.6
PyYAML>=6.0.2
Pillow>=10.2.0
